treat view finding the same way blade does with dot notation.
Move all internal views into a framework "namespace"
