+++ resources/views/framework/errors/500-admin.php
