+++ resources/views/framework/redirect-protection.php
