+++ tests/fixtures/views/subdirectory/nested-view.php
@@ -1,0 +1,5 @@
+<?php
+
+declare(strict_types=1);
+
+echo "Nested View";
